19 Спринт: ДЗ. Экран Settings переведён на Fragment.
