+++ app/src/main/AndroidManifest.xml
@@ -34,11 +34,6 @@
             android:exported="false" />
 
         <activity
-            android:name=".presentation.settings.activity.SettingsActivity"
-            android:exported="false"
-            android:theme="@style/Base.Theme.PlaylistMaker.Settings" />
-
-        <activity
             android:name=".presentation.main.MainActivity"
             android:exported="true"
             android:theme="@style/Base.Theme.PlaylistMaker.Main">
