+++ app/src/main/java/com/practicum/playlistmaker/di/ViewModelModule.kt
@@ -4,7 +4,7 @@ import com.practicum.playlistmaker.presentation.library.LibraryFavouriteViewMode
 import com.practicum.playlistmaker.presentation.library.LibraryPlaylistViewModel
 import com.practicum.playlistmaker.presentation.player.view_model.PlayerViewModel
 import com.practicum.playlistmaker.presentation.search.view_model.SearchViewModel
-import com.practicum.playlistmaker.presentation.settings.view_model.SettingsViewModel
+import com.practicum.playlistmaker.presentation.settings.SettingsViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
